+++ src/main/java/ch/uzh/ifi/hase/soprafs23/repository/UserRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository("userRepository")
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
-    Optional<User> findById(Long id);
+    Optional<User> findById(int id);
     User findByToken(String token);
 
     User findByUsername(String username);
